requirements.txt
# Core Django
Django==5.1.2
djangorestframework==3.15.2

# Authentication & Authorization
django-allauth==0.57.0
djangorestframework-simplejwt==5.3.1

# API Documentation
drf-spectacular==0.27.2

# Frontend Enhancements
django-htmx==1.18.1

# CORS
django-cors-headers==4.4.0

# Database
psycopg2-binary==2.9.10

# File Storage (AWS S3)
boto3==1.35.20
django-storages==1.14.2

# Email
django-anymail==11.1

# Task Queue
celery==5.3.6
redis==5.0.6

# Internationalization
django-rosetta==0.9.9

# Rich Text Editor
django-ckeditor==6.7.1

# Image Processing
Pillow==10.4.0

# Environment Variables
python-decouple==3.8

# Code Quality
black==24.3.0
ruff==0.5.1
mypy==1.10.0
pre-commit==3.7.0

# Testing
pytest==8.2.0
pytest-django==4.8.0
pytest-cov==5.0.0
factory-boy==3.3.0

# Payments
stripe==11.0.0

# Date/Time Utilities
python-dateutil==2.9.0.post0

# CSV/Excel Export
openpyxl==3.1.5

# PDF Generation
reportlab==4.2.0

# Security
django-ratelimit==4.1.0

# API Rate Limiting
django-ratelimit==4.1.0

# Real-time Communication
channels==4.0.0
channels-redis==4.2.0
daphne==4.0.0

